Create core.magics.code according to new API.
